// src/modules/interface.ts
export interface headerItemsString {
    id: string
    title: string
}
export interface heroIconsString {
    id: string
    icon: string
}
export interface CardItemsString {
    id: string
    logo: string
    title: string
    subTitle: string
    text: string
    iconOne: string
    iconTwo: string
}
export interface trustedImagesString {
    id: string
    icon: string
}
export interface buttonType {
    label: string
    border: string
    backgound: string
    hover: string
    textColor: string
}